Update version of sassc to 3.3.0/3.3.2
This was lagging behind what is available elsewhere and causing
conflicts between my local dev environment and 'production'.

This commit introduces a Vagrant script that will build sassc and drop
it into the support directory, so in future it should just be a digit
bump, run the script and done.